requirements.txt
gunicorn==20.0.4
psycopg2==2.9.3
psycopg2-binary==2.9.4
vkbottle==4.3.10
python-dotenv==0.21.0
vkcoin==2.0.4
requests==2.28.1